+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ### 0.0.3-pre
-*
+* Stop blinking when off is called
 
 ### 0.0.2
 * Add blink method